data-pipeline/requirements.txt
beautifulsoup4==4.12.3
docling==2.61.2
pymongo==4.8.0
pinecone-client==4.1.0
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.5
PyPDF2==3.0.1
openai==1.55.3
pandas==2.2.3
